{"schema":"decodex/commit/1","summary":"Split MCP resource context modules","authority":"XY-1115"}
