+++ apps/decodex/src/mcp/resources/context.rs
@@ -1,380 +1,4 @@
-use std::{
-	fs,
-	io::ErrorKind,
-	path::{Path, PathBuf},
-};
-
-use serde_json::{self, Value};
-
-use crate::orchestrator;
-
-use super::{
-	super::{
-		DEFAULT_MCP_STATUS_LIMIT, McpContext, McpError,
-		autonomy_resources::{
-			mcp_autonomy_current_objective_resource, mcp_autonomy_evidence_resource,
-			mcp_autonomy_objective_version_resource, mcp_autonomy_project_resource,
-			mcp_autonomy_proposal_resource, mcp_autonomy_proposals_resource,
-			mcp_autonomy_signal_resource, mcp_autonomy_signals_resource,
-		},
-		observability::{
-			mcp_activity_tail_resource, mcp_pr_review_state_resource,
-			mcp_public_lane_control_readback_resource, mcp_public_lane_inspect_resource,
-			mcp_run_resource, mcp_status_live_resource, sanitize_mcp_observability_value,
-		},
-		safe_autonomy_record_identifier, safe_runtime_identifier,
-	},
-	DECISION_CONTRACTS_HOST, DOCS_HOST, PROJECTS_HOST, RESEARCH_HOST,
-	types::{McpResource, ResourceContent, ResourceUri},
-};
-
-impl McpContext {
-	pub(super) fn docs_resources(&self) -> crate::prelude::Result<Vec<McpResource>, McpError> {
-		let mut resources = Vec::new();
-
-		push_file_resource(
-			&mut resources,
-			self.repo_root.join("docs/index.md"),
-			"decodex://docs/index",
-			"Documentation index",
-			"Checked-in Decodex documentation router.",
-		);
-		push_file_resource(
-			&mut resources,
-			self.repo_root.join("docs/policy.md"),
-			"decodex://docs/policy",
-			"Documentation policy",
-			"Checked-in Decodex documentation policy.",
-		);
-
-		for lane in ["spec", "runbook", "reference", "decisions", "research"] {
-			let docs_dir = self.repo_root.join("docs").join(lane);
-
-			for entry in read_sorted_dir(&docs_dir)? {
-				let Some(stem) = markdown_stem(&entry) else {
-					continue;
-				};
-
-				resources.push(McpResource::markdown(
-					format!("decodex://docs/{lane}/{stem}"),
-					format!("docs/{lane}/{stem}.md"),
-					"Checked-in Decodex documentation resource.",
-				));
-			}
-		}
-		for entry in read_sorted_dir(&self.repo_root.join("docs/research"))? {
-			let Some(stem) = markdown_stem(&entry) else {
-				continue;
-			};
-
-			resources.push(McpResource::markdown(
-				format!("decodex://research/{stem}"),
-				format!("docs/research/{stem}.md"),
-				"Checked-in Markdown Research Contract concept.",
-			));
-		}
-
-		Ok(resources)
-	}
-
-	pub(super) fn decision_contract_resources(
-		&self,
-	) -> crate::prelude::Result<Vec<McpResource>, McpError> {
-		let Some(project_id) = self.project_id.as_deref() else {
-			return Ok(Vec::new());
-		};
-		let Some(state_store) = self.state_store.as_ref() else {
-			return Ok(Vec::new());
-		};
-		let records = state_store
-			.list_decision_contracts_for_project(project_id)
-			.map_err(McpError::internal)?;
-
-		Ok(records
-			.into_iter()
-			.map(|record| {
-				McpResource::json(
-					format!("decodex://decision-contracts/{}", record.contract_id()),
-					format!("Decision Contract {}", record.contract_id()),
-					"Read-only local runtime Decision Contract readback.",
-				)
-			})
-			.collect())
-	}
-
-	pub(super) fn read_resource(
-		&self,
-		uri: &str,
-	) -> crate::prelude::Result<ResourceContent, McpError> {
-		let resource_uri = ResourceUri::parse(uri)?;
-
-		match resource_uri.host.as_str() {
-			DOCS_HOST => self.read_docs_resource(&resource_uri),
-			RESEARCH_HOST => self.read_research_resource(&resource_uri),
-			DECISION_CONTRACTS_HOST => self.read_decision_contract_resource(&resource_uri),
-			PROJECTS_HOST => self.read_project_resource(&resource_uri),
-			_ => Err(McpError::resource_not_found()),
-		}
-	}
-
-	fn read_docs_resource(
-		&self,
-		uri: &ResourceUri,
-	) -> crate::prelude::Result<ResourceContent, McpError> {
-		let path = match uri.segments.as_slice() {
-			[segment] if segment == "index" => self.repo_root.join("docs/index.md"),
-			[segment] if segment == "policy" => self.repo_root.join("docs/policy.md"),
-			[lane, topic] if docs_lane_allowed(lane) && safe_resource_stem(topic) =>
-				self.repo_root.join("docs").join(lane).join(format!("{topic}.md")),
-			_ => return Err(McpError::resource_not_found()),
-		};
-
-		read_file_resource(&uri.raw, path, "text/markdown")
-	}
-
-	fn read_research_resource(
-		&self,
-		uri: &ResourceUri,
-	) -> crate::prelude::Result<ResourceContent, McpError> {
-		let [concept] = uri.segments.as_slice() else {
-			return Err(McpError::resource_not_found());
-		};
-
-		if !safe_resource_stem(concept) {
-			return Err(McpError::resource_not_found());
-		}
-
-		read_file_resource(
-			&uri.raw,
-			self.repo_root.join("docs/research").join(format!("{concept}.md")),
-			"text/markdown",
-		)
-	}
-
-	fn read_decision_contract_resource(
-		&self,
-		uri: &ResourceUri,
-	) -> crate::prelude::Result<ResourceContent, McpError> {
-		let [contract_id] = uri.segments.as_slice() else {
-			return Err(McpError::resource_not_found());
-		};
-
-		if !safe_runtime_identifier(contract_id) {
-			return Err(McpError::resource_not_found());
-		}
-
-		let Some(project_id) = self.project_id.as_deref() else {
-			return Err(McpError::resource_not_found());
-		};
-		let Some(state_store) = self.state_store.as_ref() else {
-			return Err(McpError::resource_not_found());
-		};
-		let Some(record) =
-			state_store.decision_contract(project_id, contract_id).map_err(McpError::internal)?
-		else {
-			return Err(McpError::resource_not_found());
-		};
-		let mut value = serde_json::json!({
-			"schema": "decodex.mcp.decision_contract_resource/1",
-			"project_id": record.project_id(),
-			"source_issue_id": record.source_issue_id(),
-			"status": record.status(),
-			"created_at": record.created_at(),
-			"updated_at": record.updated_at(),
-			"decision_contract": record.contract()
-		});
-
-		sanitize_mcp_observability_value(&mut value);
-
-		ResourceContent::json(&uri.raw, value)
-	}
-
-	fn read_project_resource(
-		&self,
-		uri: &ResourceUri,
-	) -> crate::prelude::Result<ResourceContent, McpError> {
-		let [project_id, resource_kind, rest @ ..] = uri.segments.as_slice() else {
-			return Err(McpError::resource_not_found());
-		};
-
-		if Some(project_id.as_str()) != self.project_id.as_deref() {
-			return Err(McpError::resource_not_found());
-		}
-		if resource_kind == "autonomy" {
-			let value = self.read_autonomy_project_resource(project_id, rest)?;
-
-			return ResourceContent::mcp_observability_json(&uri.raw, value);
-		}
-
-		let Some(config_path) = self.config_path.as_deref() else {
-			return Err(McpError::resource_not_found());
-		};
-		let value = match (resource_kind.as_str(), rest) {
-			("status", []) =>
-				orchestrator::build_mcp_status_resource(Some(config_path), DEFAULT_MCP_STATUS_LIMIT)
-					.map_err(McpError::internal),
-			("status_live", []) =>
-				orchestrator::build_mcp_status_resource(Some(config_path), DEFAULT_MCP_STATUS_LIMIT)
-					.map(mcp_status_live_resource)
-					.map_err(McpError::internal),
-			("activity_tail", []) =>
-				orchestrator::build_mcp_status_resource(Some(config_path), DEFAULT_MCP_STATUS_LIMIT)
-					.map(mcp_activity_tail_resource)
-					.map_err(McpError::internal),
-			("lane-control", []) => orchestrator::build_mcp_lane_control_resource(
-				Some(config_path),
-				None,
-				None,
-				DEFAULT_MCP_STATUS_LIMIT,
-			)
-			.map(mcp_public_lane_control_readback_resource)
-			.map_err(McpError::internal),
-			("lane-control", [issue]) if safe_runtime_identifier(issue) =>
-				orchestrator::build_mcp_lane_control_resource(
-					Some(config_path),
-					Some(issue),
-					None,
-					DEFAULT_MCP_STATUS_LIMIT,
-				)
-				.map(mcp_public_lane_inspect_resource)
-				.map_err(McpError::internal),
-			("lane_inspect", [issue]) if safe_runtime_identifier(issue) =>
-				orchestrator::build_mcp_lane_control_resource(
-					Some(config_path),
-					Some(issue),
-					None,
-					DEFAULT_MCP_STATUS_LIMIT,
-				)
-				.map(mcp_public_lane_inspect_resource)
-				.map_err(McpError::internal),
-			("runs", [run_id, resource])
-				if resource == "events" && safe_runtime_identifier(run_id) =>
-				orchestrator::build_mcp_status_resource(Some(config_path), DEFAULT_MCP_STATUS_LIMIT)
-					.map_err(McpError::internal)
-					.and_then(|snapshot| mcp_run_resource(&snapshot, run_id, "events")),
-			("runs", [run_id, resource])
-				if resource == "protocol_activity" && safe_runtime_identifier(run_id) =>
-				orchestrator::build_mcp_status_resource(Some(config_path), DEFAULT_MCP_STATUS_LIMIT)
-					.map_err(McpError::internal)
-					.and_then(|snapshot| mcp_run_resource(&snapshot, run_id, "protocol_activity")),
-			("runs", [run_id, resource])
-				if resource == "child_agent_activity" && safe_runtime_identifier(run_id) =>
-				orchestrator::build_mcp_status_resource(Some(config_path), DEFAULT_MCP_STATUS_LIMIT)
-					.map_err(McpError::internal)
-					.and_then(|snapshot| {
-						mcp_run_resource(&snapshot, run_id, "child_agent_activity")
-					}),
-			("runs", [run_id, resource])
-				if resource == "progress_diagnostics" && safe_runtime_identifier(run_id) =>
-				orchestrator::build_mcp_status_resource(Some(config_path), DEFAULT_MCP_STATUS_LIMIT)
-					.map_err(McpError::internal)
-					.and_then(|snapshot| {
-						mcp_run_resource(&snapshot, run_id, "progress_diagnostics")
-					}),
-			("pr_review_state", []) =>
-				orchestrator::build_mcp_status_resource(Some(config_path), DEFAULT_MCP_STATUS_LIMIT)
-					.map(mcp_pr_review_state_resource)
-					.map_err(McpError::internal),
-			_ => return Err(McpError::resource_not_found()),
-		}?;
-
-		ResourceContent::mcp_observability_json(&uri.raw, value)
-	}
-
-	fn read_autonomy_project_resource(
-		&self,
-		project_id: &str,
-		rest: &[String],
-	) -> crate::prelude::Result<Value, McpError> {
-		let Some(state_store) = self.state_store.as_ref() else {
-			return Err(McpError::resource_not_found());
-		};
-
-		match rest {
-			[] => mcp_autonomy_project_resource(state_store, project_id),
-			[resource] if resource == "signals" =>
-				mcp_autonomy_signals_resource(state_store, project_id),
-			[resource, signal_id]
-				if resource == "signals" && safe_autonomy_record_identifier(signal_id) =>
-				mcp_autonomy_signal_resource(state_store, project_id, signal_id),
-			[resource] if resource == "proposals" =>
-				mcp_autonomy_proposals_resource(state_store, project_id),
-			[resource, proposal_id]
-				if resource == "proposals" && safe_autonomy_record_identifier(proposal_id) =>
-				mcp_autonomy_proposal_resource(state_store, project_id, proposal_id),
-			[resource] if resource == "evidence" =>
-				mcp_autonomy_evidence_resource(state_store, project_id),
-			[resource, objective_id, selector]
-				if resource == "objectives"
-					&& safe_runtime_identifier(objective_id)
-					&& selector == "current" =>
-				mcp_autonomy_current_objective_resource(state_store, project_id, objective_id),
-			[resource, objective_id, version]
-				if resource == "objectives" && safe_runtime_identifier(objective_id) =>
-				mcp_autonomy_objective_version_resource(
-					state_store,
-					project_id,
-					objective_id,
-					version,
-				),
-			_ => Err(McpError::resource_not_found()),
-		}
-	}
-}
-
-fn push_file_resource(
-	resources: &mut Vec<McpResource>,
-	path: PathBuf,
-	uri: &str,
-	name: &str,
-	description: &str,
-) {
-	if path.is_file() {
-		resources.push(McpResource::markdown(uri, name, description));
-	}
-}
-
-fn read_sorted_dir(path: &Path) -> crate::prelude::Result<Vec<PathBuf>, McpError> {
-	let entries = match fs::read_dir(path) {
-		Ok(entries) => entries,
-		Err(error) if error.kind() == ErrorKind::NotFound => return Ok(Vec::new()),
-		Err(error) => return Err(McpError::internal(error)),
-	};
-	let mut paths = entries
-		.map(|entry| entry.map(|entry| entry.path()).map_err(McpError::internal))
-		.collect::<crate::prelude::Result<Vec<_>, _>>()?;
-
-	paths.sort();
-
-	Ok(paths)
-}
-
-fn markdown_stem(path: &Path) -> Option<String> {
-	if path.extension().and_then(|extension| extension.to_str()) != Some("md") {
-		return None;
-	}
-
-	path.file_stem().and_then(|stem| stem.to_str()).map(str::to_owned)
-}
-
-fn read_file_resource(
-	uri: &str,
-	path: PathBuf,
-	mime_type: &str,
-) -> crate::prelude::Result<ResourceContent, McpError> {
-	let text = fs::read_to_string(path).map_err(|error| match error.kind() {
-		ErrorKind::NotFound => McpError::resource_not_found(),
-		_ => McpError::internal(error),
-	})?;
-
-	Ok(ResourceContent { uri: uri.to_owned(), mime_type: mime_type.to_owned(), text })
-}
-
-fn docs_lane_allowed(lane: &str) -> bool {
-	matches!(lane, "spec" | "runbook" | "reference" | "decisions" | "research")
-}
-
-fn safe_resource_stem(value: &str) -> bool {
-	!value.is_empty()
-		&& value.bytes().all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_'))
-}
+mod decision_contracts;
+mod docs;
+mod project;
+mod router;
