+++ apps/decodex/src/mcp/resources/context/decision_contracts.rs
@@ -1,0 +1,72 @@
+use crate::{
+	mcp::{
+		self, McpContext, McpError, observability,
+		resources::types::{McpResource, ResourceContent, ResourceUri},
+	},
+	prelude::Result,
+};
+
+impl McpContext {
+	pub(in crate::mcp::resources) fn decision_contract_resources(
+		&self,
+	) -> Result<Vec<McpResource>, McpError> {
+		let Some(project_id) = self.project_id.as_deref() else {
+			return Ok(Vec::new());
+		};
+		let Some(state_store) = self.state_store.as_ref() else {
+			return Ok(Vec::new());
+		};
+		let records = state_store
+			.list_decision_contracts_for_project(project_id)
+			.map_err(McpError::internal)?;
+
+		Ok(records
+			.into_iter()
+			.map(|record| {
+				McpResource::json(
+					format!("decodex://decision-contracts/{}", record.contract_id()),
+					format!("Decision Contract {}", record.contract_id()),
+					"Read-only local runtime Decision Contract readback.",
+				)
+			})
+			.collect())
+	}
+
+	pub(super) fn read_decision_contract_resource(
+		&self,
+		uri: &ResourceUri,
+	) -> Result<ResourceContent, McpError> {
+		let [contract_id] = uri.segments.as_slice() else {
+			return Err(McpError::resource_not_found());
+		};
+
+		if !mcp::safe_runtime_identifier(contract_id) {
+			return Err(McpError::resource_not_found());
+		}
+
+		let Some(project_id) = self.project_id.as_deref() else {
+			return Err(McpError::resource_not_found());
+		};
+		let Some(state_store) = self.state_store.as_ref() else {
+			return Err(McpError::resource_not_found());
+		};
+		let Some(record) =
+			state_store.decision_contract(project_id, contract_id).map_err(McpError::internal)?
+		else {
+			return Err(McpError::resource_not_found());
+		};
+		let mut value = serde_json::json!({
+			"schema": "decodex.mcp.decision_contract_resource/1",
+			"project_id": record.project_id(),
+			"source_issue_id": record.source_issue_id(),
+			"status": record.status(),
+			"created_at": record.created_at(),
+			"updated_at": record.updated_at(),
+			"decision_contract": record.contract()
+		});
+
+		observability::sanitize_mcp_observability_value(&mut value);
+
+		ResourceContent::json(&uri.raw, value)
+	}
+}
